refactor: separate sort data logic
